requirements.txt
Django==3.1.3
django-allauth==0.41.0
django-bootstrap4==2.2.0
django-filter==2.2.0
django-hosts==4.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
django-tenant-schemas==1.10.0
django-tenant-users==0.3.12
django-tenants==3.2.1
djangorestframework==3.11.0
djoser==2.0.5
gunicorn==20.0.4
paystackapi==2.0.0
psycopg2-binary==2.8.6
